城市切换弹窗与界面刷新

+++ HDMiusiSpace/Section/D_Exhibition/View/Cell/HDSSL_dExhibitionCell.swift
@@ -42,7 +42,7 @@ class HDSSL_dExhibitionCell: UITableViewCell {
             cellTitle.text = model?.title
             cell_locaAndPrice.text = model?.address
             cell_number.text = model!.star
-            let star: Int! = Int(model!.star ?? "0")
+            let star: Float! = Float(model!.star ?? "0")
             var imgStr = ""
             if star < 2 {
                 imgStr = "exhibitionCmt_1_5"
